dev: ds1307: add Epson RX-8035 support
This I2C RTC has a similar time register layout to the ds1307 family,
but with some minor differences in bit positions, polarity and control
registers.

There are also other features in this RTC like event (GPIO) detection,
alarm, timestamp and scratch/RAM register which are not implemented.
These may justify a separate driver in the future.

This driver has been tested to be interoperable with the Linux
rtc-rx8025 driver on the Traverse Technologies Ten64 board.